Refactor document selector into provider and menu
